Modify to make it work

+++ TextRPG/TextRPG/Random.cpp
@@ -10,9 +10,9 @@ CRandom::~CRandom()
 
 }
 
-int CRandom::getInt(int start, int end)
+int CRandom::getInt(size_t start, size_t end)
 {
-	uniform_int_distribution<int> dist(start, end);
+	uniform_int_distribution<size_t> dist(start, end);
 
 	return dist(this->gen);
 }
